% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlates.R
\docType{data}
\name{correlates24}
\alias{correlates24}
\title{The Correlates of State Policy Project Dataset (IPPSR)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6171 rows and 2279 columns.
}
\usage{
correlates24
}
\description{
The Correlates of State Policy Project dataset assembled by Marty P. Jordan and Matt Grossmann (2020) <http://ippsr.msu.edu/public-policy/correlates-state-policy>. Each observation is a state-year.
}
\details{
@name correlates24

 @docType data

 @usage data(correlates24)


 @keywords datasets
}
\keyword{datasets}
